backend/requirements.txt
asgiref==3.7.1
async-generator==1.10    
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
beautifulsoup4==4.12.2   
bs4==0.0.1
certifi==2023.5.7        
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0    
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0       
croniter==1.3.14
cryptography==40.0.2     
daphne==4.0.0
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
django-cron==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
fake-useragent==1.1.3
filelock==3.12.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
lxml==4.9.2
msgpack==1.0.5
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.3
outcome==1.2.0
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.5.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
pyOpenSSL==23.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
requests==2.31.0
rest-framework-simplejwt==0.0.2
routing==0.2.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.1
tzdata==2023.3
Unidecode==1.3.6
urllib3==2.0.2
wsproto==1.2.0
zope.interface==6.0